+++ examples/with-next-app-i18n/app/[locale]/providers.tsx
@@ -20,7 +20,6 @@ import {
   optimism,
   polygon,
   sepolia,
-  zora,
 } from 'wagmi/chains';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
@@ -35,7 +34,6 @@ const config = getDefaultConfig({
     optimism,
     arbitrum,
     base,
-    zora,
     ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === 'true' ? [sepolia] : []),
   ],
   wallets: [